Finite Mixture Analysis table from Godfrey et al. (1993)

Values of k for a given sample size for the calculation of distribution means for a Gaussian finite mixture using equations in Godfrey et al. (1993). k gives the mean expected number of standard deviations (as measured for the whole dataset) represented in the observed data range for a given sample size.

Usage

data(FMAtable)

Format

An object of class data.frame with 99 rows and 2 columns.

References

Godfrey LR, Lyon SK, Sutherland MR. (1993) Sexual dimorphism in large-bodied primates: The case of the subfossil lemurs. American Journal of Physical Anthropology. 90:315-334. (https://doi.org/10.1002/ajpa.1330900306)

Examples

data(FMAtable)
FMAtable
#>      n        k
#> 1    2 1.283800
#> 2    3 1.692570
#> 3    4 2.058750
#> 4    5 2.325930
#> 5    6 2.534410
#> 6    7 2.704360
#> 7    8 2.847200
#> 8    9 2.970030
#> 9   10 3.077510
#> 10  11 3.172870
#> 11  12 3.258460
#> 12  13 3.335980
#> 13  14 3.406760
#> 14  15 3.471830
#> 15  16 3.531980
#> 16  17 3.587880
#> 17  18 3.640060
#> 18  19 3.688960
#> 19  20 3.734950
#> 20  21 3.778340
#> 21  22 3.819380
#> 22  23 3.858320
#> 23  24 3.895350
#> 24  25 3.930630
#> 25  26 3.964320
#> 26  27 3.996540
#> 27  28 4.027410
#> 28  29 4.057040
#> 29  30 4.085520
#> 30  31 4.110747
#> 31  32 4.136129
#> 32  33 4.161668
#> 33  34 4.187365
#> 34  35 4.213220
#> 35  36 4.234669
#> 36  37 4.256226
#> 37  38 4.277894
#> 38  39 4.299671
#> 39  40 4.321560
#> 40  41 4.340175
#> 41  42 4.358870
#> 42  43 4.377646
#> 43  44 4.396502
#> 44  45 4.415440
#> 45  46 4.431859
#> 46  47 4.448340
#> 47  48 4.464882
#> 48  49 4.481485
#> 49  50 4.498150
#> 50  51 4.512818
#> 51  52 4.527534
#> 52  53 4.542298
#> 53  54 4.557110
#> 54  55 4.571970
#> 55  56 4.585211
#> 56  57 4.598491
#> 57  58 4.611808
#> 58  59 4.625165
#> 59  60 4.638560
#> 60  61 4.650617
#> 61  62 4.662706
#> 62  63 4.674826
#> 63  64 4.686977
#> 64  65 4.699160
#> 65  66 4.710220
#> 66  67 4.721306
#> 67  68 4.732418
#> 68  69 4.743556
#> 69  70 4.754720
#> 70  71 4.764928
#> 71  72 4.775158
#> 72  73 4.785410
#> 73  74 4.795684
#> 74  75 4.805980
#> 75  76 4.815457
#> 76  77 4.824952
#> 77  78 4.834466
#> 78  79 4.843998
#> 79  80 4.853550
#> 80  81 4.862386
#> 81  82 4.871238
#> 82  83 4.880106
#> 83  84 4.888990
#> 84  85 4.897890
#> 85  86 4.906164
#> 86  87 4.914452
#> 87  88 4.922754
#> 88  89 4.931070
#> 89  90 4.939400
#> 90  91 4.947177
#> 91  92 4.954967
#> 92  93 4.962769
#> 93  94 4.970583
#> 94  95 4.978410
#> 95  96 4.985744
#> 96  97 4.993090
#> 97  98 5.000446
#> 98  99 5.007812
#> 99 100 5.015190