Long bone articular dimensions for two simulated fossil hominoid samples with missing data

Postcranial metric data collected for two "fauxil" species. Fauxil sp. 1 individuals are extant gorillas from the apelimbart dataset with some data removed, and Fauxil sp. 2 individuals are modern humans from the apelimbart dataset with some data removed. All data are from Gordon (2025a).

Usage

fauxil

Format

A data frame with 35 rows and 13 variables:

Species

factor Species name, with two possible levels (in decreasing order of postcranial dimorphism: "Fauxil sp. 1" and "Fauxil sp. 2")

Museum

factor Collection housing each specimen

Collection.ID

character Unique specimen identifier within each collection

FHSI

numeric Femoral head superoinferior diameter: the maximum superoinferior diameter of the femoral head.

TPML

numeric Tibial plateau mediolateral width: the maximum mediolateral width of the articular surface of the tibial plateau.

TPMAP

numeric Tibial plateau medial condyle anteroposterior length: the maximum anteroposterior length of the articular surface of the medial condyle of the tibial plateau.

TPLAP

numeric Tibial plateau lateral condyle anteroposterior length: the maximum anteroposterior length of the articular surface of the lateral condyle of the tibial plateau.

HHMaj

numeric Humeral head major axis diameter: treating the articular surface of the humeral head as a partial oblate spheroid, this is the length of the major axis passing through the oblate spheroid.

HHMin

numeric Humeral head minor axis diameter: the maximum width of the articular surface of the humeral head perpendicular to the major axis.

RHMaj

numeric Radial head major axis diameter: treating the radial head in proximal view as an ellipse, this is the length of the major axis passing through the ellipse.

RHMin

numeric Radial head minor axis diameter: the maximum width of the radial head perpendicular to the major axis.

RDAP

numeric Distal radius anteroposterior width: the distance between the anterior and posterior extents of the boundary between the lunate and scaphoid facets of the distal articular surface of the radius.

RDML

numeric Distal radius mediolateral breadth: the maximum width of the distal articular surface of the radius when the medial point of this dimension is constrained to the midpoint of the curve of the articulation with the distal ulna.

References

Gordon AD. (2025a) Interpreting statistical significance in hominin dimorphism: Power and Type I error rates for resampling tests of univariate and missing-data multivariate size dimorphism estimation methods in the fossil record. Journal of Human Evolution. 199:103630. (https://doi.org/10.1016/j.jhevol.2024.103630)

See also

Examples

data(fauxil)
fauxil
#>                    Species Museum Collection.ID  FHSI  TPML TPMAP TPLAP HHMaj
#> PCM Gg-M877   Fauxil sp. 1    PCM       Gg-M877    NA    NA 40.71    NA    NA
#> PCM Gg-Z6-33  Fauxil sp. 1    PCM      Gg-Z6-33    NA 69.42    NA    NA 46.18
#> CMNH HTB 1997 Fauxil sp. 1   CMNH      HTB 1997    NA    NA    NA    NA 50.86
#> CMNH HTB 1710 Fauxil sp. 1   CMNH      HTB 1710 39.85 66.34 41.42    NA    NA
#> CMNH HTB 1846 Fauxil sp. 1   CMNH      HTB 1846    NA    NA    NA 32.39 49.47
#> CMNH HTB 1851 Fauxil sp. 1   CMNH      HTB 1851    NA    NA    NA 37.26    NA
#> ZSM 1954/0201 Fauxil sp. 1    ZSM     1954/0201    NA    NA    NA    NA    NA
#> MNHN 1982-56  Fauxil sp. 1   MNHN       1982-56    NA    NA    NA    NA    NA
#> CMNH HTB 1729 Fauxil sp. 1   CMNH      HTB 1729 47.93    NA    NA    NA    NA
#> PCM Gg-C1-105 Fauxil sp. 1    PCM     Gg-C1-105    NA    NA    NA    NA    NA
#> CMNH HTB 1994 Fauxil sp. 1   CMNH      HTB 1994 49.68    NA    NA 46.94 60.95
#> CMNH HTB 1732 Fauxil sp. 1   CMNH      HTB 1732    NA    NA    NA 40.46    NA
#> CMNH HTB 1954 Fauxil sp. 1   CMNH      HTB 1954    NA    NA    NA    NA    NA
#> CMNH HTB 1797 Fauxil sp. 1   CMNH      HTB 1797    NA    NA    NA    NA 66.85
#> MNHN 1931-657 Fauxil sp. 1   MNHN      1931-657 53.17 86.29    NA    NA    NA
#> CMNH HTB 3400 Fauxil sp. 1   CMNH      HTB 3400    NA 93.74    NA    NA    NA
#> CMNH HTH 1709 Fauxil sp. 2   CMNH      HTH 1709 39.73    NA    NA    NA    NA
#> CMNH HTH 1779 Fauxil sp. 2   CMNH      HTH 1779    NA    NA    NA    NA 40.01
#> CMNH HTH 0221 Fauxil sp. 2   CMNH      HTH 0221    NA    NA    NA 36.78    NA
#> CMNH HTH 1270 Fauxil sp. 2   CMNH      HTH 1270 41.80    NA 46.30    NA 40.16
#> CMNH HTH 2056 Fauxil sp. 2   CMNH      HTH 2056 43.86    NA 45.85    NA    NA
#> CMNH HTH 2761 Fauxil sp. 2   CMNH      HTH 2761    NA 72.32    NA    NA 41.46
#> CMNH HTH 0727 Fauxil sp. 2   CMNH      HTH 0727    NA    NA 50.17    NA    NA
#> CMNH HTH 1755 Fauxil sp. 2   CMNH      HTH 1755 44.91    NA    NA    NA    NA
#> CMNH HTH 2428 Fauxil sp. 2   CMNH      HTH 2428 45.32    NA    NA    NA 45.66
#> CMNH HTH 1685 Fauxil sp. 2   CMNH      HTH 1685    NA    NA 47.52    NA 46.81
#> CMNH HTH 0383 Fauxil sp. 2   CMNH      HTH 0383    NA    NA    NA    NA    NA
#> CMNH HTH 0280 Fauxil sp. 2   CMNH      HTH 0280 47.46    NA 44.99    NA    NA
#> CMNH HTH 0922 Fauxil sp. 2   CMNH      HTH 0922 47.48    NA    NA    NA    NA
#> CMNH HTH 0878 Fauxil sp. 2   CMNH      HTH 0878    NA    NA    NA 44.17    NA
#> CMNH HTH 1398 Fauxil sp. 2   CMNH      HTH 1398    NA    NA    NA 42.80    NA
#> CMNH HTH 2584 Fauxil sp. 2   CMNH      HTH 2584    NA    NA    NA    NA    NA
#> CMNH HTH 1389 Fauxil sp. 2   CMNH      HTH 1389 49.30    NA    NA    NA    NA
#> CMNH HTH 0720 Fauxil sp. 2   CMNH      HTH 0720    NA 76.92    NA    NA    NA
#> CMNH HTH 0712 Fauxil sp. 2   CMNH      HTH 0712    NA    NA    NA 44.53    NA
#>               HHMin RHMaj RHMin  RDAP  RDML
#> PCM Gg-M877      NA    NA 25.67    NA    NA
#> PCM Gg-Z6-33     NA    NA    NA    NA 28.87
#> CMNH HTB 1997 42.31    NA    NA    NA    NA
#> CMNH HTB 1710 43.04    NA 24.72 19.63    NA
#> CMNH HTB 1846 44.10    NA    NA    NA    NA
#> CMNH HTB 1851    NA    NA    NA 20.79    NA
#> ZSM 1954/0201    NA 26.83 26.00    NA    NA
#> MNHN 1982-56     NA    NA 30.50    NA    NA
#> CMNH HTB 1729    NA    NA    NA 26.68    NA
#> PCM Gg-C1-105 59.26    NA    NA    NA 36.06
#> CMNH HTB 1994    NA 34.65    NA    NA    NA
#> CMNH HTB 1732 56.03    NA    NA    NA    NA
#> CMNH HTB 1954 57.52    NA    NA    NA    NA
#> CMNH HTB 1797    NA    NA    NA    NA    NA
#> MNHN 1931-657    NA    NA    NA    NA    NA
#> CMNH HTB 3400    NA    NA    NA    NA    NA
#> CMNH HTH 1709    NA    NA    NA 13.82    NA
#> CMNH HTH 1779    NA    NA    NA    NA 26.13
#> CMNH HTH 0221    NA    NA    NA    NA    NA
#> CMNH HTH 1270    NA    NA    NA    NA    NA
#> CMNH HTH 2056    NA    NA    NA    NA    NA
#> CMNH HTH 2761    NA    NA    NA 13.79    NA
#> CMNH HTH 0727    NA    NA    NA    NA    NA
#> CMNH HTH 1755    NA    NA    NA    NA 28.89
#> CMNH HTH 2428    NA 23.39    NA 14.53    NA
#> CMNH HTH 1685    NA 26.30    NA    NA    NA
#> CMNH HTH 0383    NA 21.04    NA    NA 27.48
#> CMNH HTH 0280    NA    NA    NA    NA    NA
#> CMNH HTH 0922    NA    NA    NA    NA    NA
#> CMNH HTH 0878    NA    NA    NA    NA    NA
#> CMNH HTH 1398 44.71    NA    NA    NA    NA
#> CMNH HTH 2584    NA    NA    NA    NA 30.52
#> CMNH HTH 1389    NA    NA    NA    NA    NA
#> CMNH HTH 0720    NA    NA    NA    NA    NA
#> CMNH HTH 0712 47.62    NA    NA    NA 32.15